build.ext.nblist {prabclus}	R Documentation

Internal: generates neighborhood list for diploid loci

Description

This is for use in alleleinit. Given a neighborhood list of individuals, a new neighborhood list is generated in which there are two entries for each individual (entry 1 and 2 refer to individual one, 3 and 4 to individual 2 and so on). Neighborhoods are preserved and additionally the two entries belonging to the same individual are marked as neighbors.

Usage

  build.ext.nblist(neighbors,n.individuals=length(neighbors))

Arguments

neighbors
list of integer vectors, where each vector contains the neighbors of an individual.

n.individuals
integer. Number of individuals.

Value

list with 2*n.inidividuals vectors of integers as described above.

Author(s)

Christian Hennig chrish@stats.ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche

See Also

alleleinit

Examples

  data(veronica)
  vnb <- coord2dist(coordmatrix=veronica.coord[1:20,], cut=20,
    file.format="decimal2",neighbors=TRUE)
  build.ext.nblist(vnb$nblist)

[Package prabclus version 2.2-6 Index]