comp.test {prabclus}	R Documentation

Compare species clustering and species groups

Description

Tests for independence between a clustering and another grouping of species. This is simply an interface to chisq.test.

Usage

comp.test(cl,spg)

Arguments

cl
a vector of integers. Clustering of species (may be taken from prabclust).

spg
a vector of integers of the same length, groups of species.

Details

chisq.test with simulated p-value is used.

Value

Output of chisq.test.

Author(s)

Christian Hennig chrish@stats.ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche

References

Hausdorf, B. and Hennig, C. (2003) Biotic Element Analysis in Biogeography. Systematic Biology 52, 717-723.

See Also

chisq.test, prabclust.

Examples

set.seed(1234)
g1 <- c(rep(1,34),rep(2,12),rep(3,15))
g2 <- sample(3,61,replace=TRUE)
comp.test(g1,g2)

[Package prabclus version 2.2-6 Index]