con.regmat {prabclus}	R Documentation

Connected regions per species

Description

Returns a vector of the numbers of connected regions per species for a presence-absence matrix.

Usage

con.regmat(regmat, neighbors, count = FALSE)

Arguments

regmat
0-1-matrix. Columns are species, rows are regions.

neighbors
A list with a component for every region. The components are vectors of integers indicating neighboring regions. A region without neighbors (e.g., an island) should be assigned a list numeric(0).

count
logical. If TRUE, the number of the currently processed species is printed.

Details

Uses con.comp.

Value

Vector of numbers of connected regions per species.

Note

Designed for use in prabtest.

Author(s)

Christian Hennig chrish@stats.ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche

See Also

con.comp, prabtest

Examples

data(nb)
set.seed(888) 
cp <- cluspop.nb(nb, p.nb=0.1, n.species=10, clus.specs=9,
                 reg.group=1:17,vector.species=c(10))
con.regmat(cp,nb)

[Package prabclus version 2.2-6 Index]