crmatrix {prabclus}	R Documentation

Region-wise cluster membership

Description

Produces a matrix with clusters as rows and regions as columns, indicating how many species present in a region belong to the clusters

Usage

crmatrix(x,xc,percentages=FALSE)

Arguments

x
object of class prab as generated by prabinit. Presence-absence data to be analyzed.

xc
object of class prabclust or comprabclust as generated by prabclust or hprabclust. The clustering.

percentages
logical. If TRUE, the output matrix will give the proportion of species from a certain region in the cluster.

Value

A clusters time regions matrix as explained above.

Author(s)

Christian Hennig chrish@stats.ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche

Examples

  options(digits=3)
  data(kykladspecreg)
  data(nb)
  set.seed(1234)
  x <- prabinit(prabmatrix=kykladspecreg, neighborhood=nb)
  xc <- prabclust(x)

  crmatrix(x,xc)
  crmatrix(x,xc, percentages=TRUE)

[Package prabclus version 2.2-6 Index]