dicedist {prabclus}	R Documentation

Dice distance matrix

Description

Computes a distance derived from Dice's coincidence index between the columns of a 0-1-matrix.

Usage

  dicedist(regmat)

Arguments

regmat
0-1-matrix. Columns are species, rows are regions.

Details

The Dice distance between two species is 1 minus the Coincidence Index, which is (2*number of regions where both species are present)/(2*number of regions where both species are present plus number of regions where at least one species is present). This is S23 in Shi (1993).

Value

A symmetrical matrix of Dice distances.

Author(s)

Christian Hennig chrish@stats.ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche

References

Shi, G. R. (1993) Multivariate data analysis in palaeoecology and palaeobiogeography - a review. Palaeogeography, Palaeoclimatology, Palaeoecology 105, 199-234.

See Also

kulczynski,jaccard

Examples

options(digits=4)
data(kykladspecreg)
dicedist(t(kykladspecreg))

[Package prabclus version 2.2-6 Index]