distratio {prabclus}	R Documentation

Distance ratio test statistics for distance based clustering

Description

Calculates the ratio between the prop smallest and largest distances of a distance matrix.

Usage

distratio(distmat, prop = 0.25)

Arguments

distmat
symmetric distance matrix.

prop
numerical. Proportion between 0 and 1.

Details

Rounding is by floor for small and ceiling for large distances.

Value

A list with components

dr
ratio of prop smallest to prop largest distances.

lowmean
mean of prop smallest distances.

himean
mean of prop smallest distances.

prop
see above.

Author(s)

Christian Hennig chrish@stats.ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche

References

Hennig, C. and Hausdorf, B. (2004) Distance-based parametric bootstrap tests for clustering of species ranges. Computational Statistics and Data Analysis 45, 875-896.

See Also

prabtest

Examples

options(digits=4)
data(kykladspecreg)
j <- jaccard(t(kykladspecreg))
distratio(j)

[Package prabclus version 2.2-6 Index]