geo2neighbor {prabclus}	R Documentation

Neighborhood list from geographical distance

Description

Generates a neighborhood list as required by prabinit from a matrix of geographical distances.

Usage

geo2neighbor(geodist,cut=0.1*max(geodist))

Arguments

geodist
dist-object or symmetric non-negative matrix. Geographical distances between regions.

cut
non-negative numerical. All pairs of regions with distance<=cut are treated as neighbors.

Value

A list of integer vectors, giving the set of neighbors for every region.

Author(s)

Christian Hennig chrish@stats.ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche

Examples

data(waterdist)
geo2neighbor(waterdist)

[Package prabclus version 2.2-6 Index]