homogen.test {prabclus}	R Documentation

Classical distance-based test for homogeneity against clustering

Description

Classical distance-based test for homogeneity against clustering. Test statistics is number of isolated vertices in the graph of smallest distances. The homogeneity model is a random graph model where ne edges are drawn from all possible edges.

Usage

homogen.test(distmat, ne = ncol(distmat), testdist = "erdos")

Arguments

distmat
numeric symmetric distance matrix.

ne
integer. Number of edges in the data graph, corresponding to smallest distances.

testdist
string. If testdist="erdos", the test distribution is a Poisson asymptotic distibution as given by Erdos and Renyi (1960). If testdist="ling", the test distribution is exact as given by Ling (1973), which needs much more computing time.

Details

The "ling"-test is one-sided (rejection if the number of isolated vertices is too large), the "erdos"-test computes a one-sided as well as a two-sided p-value.

Value

A list with components

p
p-value for one-sided test.

p.twoside
p-value for two-sided test, only if testdist="erdos".

iv
number of isolated vertices in the data.

lambda
parameter of the Poisson test distribution, only if testdist="erdos".

distcut
largest distance value for which an edge has been drawn.

ne
see above.

Author(s)

Christian Hennig chrish@stats.ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche

References

Erdos, P. and Renyi, A. (1960) On the evolution of random graphs. Publications of the Mathematical Institute of the Hungarian Academy of Sciences 5, 17-61.

Godehardt, E. and Horsch, A. (1995) Graph-Theoretic Models for Testing the Homogeneity of Data. In Gaul, W. and Pfeifer, D. (Eds.) From Data to Knowledge, Springer, Berlin, 167-176.

Ling, R. F. (1973) A probability theory of cluster analysis. Journal of the American Statistical Association 68, 159-164.

See Also

prabtest

Examples

options(digits=4)
data(kykladspecreg)
j <- jaccard(t(kykladspecreg))
homogen.test(j, testdist="erdos")
homogen.test(j, testdist="ling")

[Package prabclus version 2.2-6 Index]