hprabclust {prabclus}	R Documentation

Clustering of species ranges from presence-absence matrices (hierarchical methods)

Description

Clusters a presence-absence matrix object by taking the 'h-cut'-partition of a hierarchical clustering and declaring all members of too small clusters as 'noise' (this gives a distance-based clustering method, which estimates the number of clusters and allows for noise/non-clustered points). Note that this is experimental. Often, the prabclust-solutions is more convincing due to higher flexibility of that method. However, hprabclust may be more stable sometimes.

Note: Data formats are described on the prabinit help page. You may also consider the example datasets kykladspecreg.dat and nb.dat. Take care of the parameter rows.are.species of prabinit.

Usage

hprabclust(prabobj, cutdist=0.4, cutout=1,
method="average", nnout=2, mdsplot=TRUE, mdsmethod="classical")

## S3 method for class 'comprabclust'
print(x, ...)

Arguments

prabobj
object of class prab as generated by prabinit. Presence-absence data to be analyzed.

cutdist
non-negative integer. Cutoff distance to determine the partition, see cutree.

cutout
non-negative integer. Points that have at most nnout distances smaller or equal than cutout are treated as noise.

method
string. Clustering method, see hclust.

nnout
non-negative integer. Members of clusters with less or equal than nnout points or that have less or equal than nnout neighbors closer than cutout are treated as noise.

mdsplot
logical. If TRUE, the cluster solution is plotted on the first two MDS dimensions, see mdsmethod.

mdsmethod
"classical", "kruskal", or "sammon". The MDS method to transform the distances to data points. "classical" indicates metric MDS by function cmdscale, "kruskal" is non-metric MDS. Note that if mdsmethod!="classical" zero distances between different objects are replaced by the minimum of the nonzero distances divided by 10 (otherwise the MDS method would produce an error). Note that mdsmethod is ignored if mdsplot=FALSE.

x
comprabclust-object as generated by hprabclus.

...
necessary for print method.

Value

hprabclust generates an object of class comprabclust. This is a list with components

clustering
vector of integers indicating the cluster memberships of the species (cutout-outliers are noise, but small clusters are allowed). Noise is coded as 0.

rclustering
vector of integers indicating the cluster memberships of the species, noise as described under nnout. Noise is coded as 0.

cutdist
see above.

method
see above.

cutout
see above.

nnout
see above.

noisen
number of points minus cutout-outliers.

symbols
vector of characters corresponding to rclustering, but estimated noise by "N".

points
numerical matrix. MDS configuration (if mdsplot=TRUE).

call
function call.

Author(s)

Christian Hennig chrish@stats.ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche

See Also

hclust, cutree, prabclust.

Examples

data(kykladspecreg)
data(nb)
data(waterdist)
x <- prabinit(prabmatrix=kykladspecreg, neighborhood=nb,
              geodist=waterdist, distance="geco")
hprabclust(x,mdsplot=FALSE)

[Package prabclus version 2.2-6 Index]