incmatrix {prabclus}	R Documentation

Nestedness matrix

Description

Computes species*species nestedness matrix and number of nestings (inclusions) from regions*species presence-absence matrix.

Usage

incmatrix(regmat)

Arguments

regmat
0-1-matrix. Columns are species, rows are regions.

Value

A list with components

m
0-1-matrix. m[i,j]=1 means that the occupied region of species j is a subset (not equal) of the region of species i.

ninc
integer. Number of strict inclusions.

neq
integer. Number of region equalities between species (not including equality between species i and i).

Author(s)

Christian Hennig chrish@stats.ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche

References

Hausdorf, B. and Hennig, C. (2003) Nestedness of nerth-west European land snail ranges as a consequence of differential immigration from Pleistocene glacial refuges. Oecologia 135, 102-109.

See Also

prabtest

Examples

data(kykladspecreg)
incmatrix(t(kykladspecreg))$ninc

[Package prabclus version 2.2-6 Index]