jaccard {prabclus}	R Documentation

Jaccard distance matrix

Description

Computes Jaccard distances between the columns of a 0-1-matrix.

Usage

jaccard(regmat)

Arguments

regmat
0-1-matrix. Columns are species, rows are regions.

Details

The Jaccard distance between two species is 1-(number of regions where both species are present)/(number of regions where at least one species is present). As a similarity coefficient, this is S22 in Shi (1993).

Thank you to Laurent Buffat for improving this function!

Value

A symmetrical matrix of Jaccard distances.

Author(s)

Christian Hennig chrish@stats.ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche

References

Shi, G. R. (1993) Multivariate data analysis in palaeoecology and palaeobiogeography - a review. Palaeogeography, Palaeoclimatology, Palaeoecology 105, 199-234.

See Also

kulczynski, dicedist

Examples

options(digits=4)
data(kykladspecreg)
jaccard(t(kykladspecreg))

[Package prabclus version 2.2-6 Index]