kykladspecreg {prabclus}	R Documentation

Snail presence-absence data from Aegean sea

Description

0-1-matrix where rows are snail species and columns are islands in the Aegean sea. An entry of 1 means that the species is present in the region.

Usage

data(kykladspecreg)

Format

A 0-1 matrix with 80 rows and 34 columns.

Details

Reads from example data file kykladspecreg.dat.

Source

B. Hausdorf and C. Hennig (2005) The influence of recent geography, palaeography and climate on the composition of the faune of the central Aegean Islands. Biological Journal of the Linnean Society 84, 785-795.

See Also

nb provides neighborhood information about the 34 islands. waterdist provides a geographical distance matrix between the islands.

Examples

data(kykladspecreg)

[Package prabclus version 2.2-6 Index]