lcomponent {prabclus}	R Documentation

Largest connectivity component

Description

Computes the size of the largest connectivity component of the graph of ncol(distmat) vertices with edges defined by the smallest ne distances.

Usage

lcomponent(distmat, ne = floor(3*ncol(distmat)/4))

Arguments

distmat
symmetric distance matrix.

ne
integer.

Value

list with components

lc
size of the largest connectivity component.

ne
see above.

Author(s)

Christian Hennig chrish@stats.ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche

References

Hennig, C. and Hausdorf, B. (2004) Distance-based parametric bootstrap tests for clustering of species ranges. Computational Statistics and Data Analysis 45, 875-896.

See Also

prabtest

Examples

data(kykladspecreg)
j <- jaccard(t(kykladspecreg))
lcomponent(j)

[Package prabclus version 2.2-6 Index]