lociplots {prabclus}	R Documentation

Visualises clusters of markers vs. species

Description

Given a clustering of individuals from prabclust (as generated in species delimitation) and a clustering of markers (for example dominant markers of genetic loci), lociplots visualises the presence of markers against the clustering of individuals and computes some statistics.

Usage

  lociplots(indclust,locclust,locprab,lcluster,
                      symbols=NULL,brightest.grey=0.8,darkest.grey=0,
                      mdsdim=1:2)

Arguments

indclust
prabclust-object. Clustering of individuals.

locclust
vector of integers. Clustering of markers/loci.

locprab
prab-object in which the markers are what the help page of prabinit refers to as "species" (i.e., reverse of what is used for species delimitation clustering; for data sets with codominant markers, such an object can be constructed by use of allele2zeroone before prabinit.)

lcluster
integer. Number of cluster in locclust for which plot and statistics are produced.

symbols
vector of plot symbols. If NULL, indclust$symbols is used.

brightest.grey
numeric between 0 and 1. Brightest grey value used in plot for individuals with smallest marker percentage, see details.

darkest.grey
numeric between 0 and 1. Darkest grey value used in plot for individuals with highest marker percentage, see details.

mdsdim
vector of two integers. The two MDS variables taken from indclust used for visualisation.

Details

Plot and statistics are based on the individual marker percentage, which is the percentage of markers present in an individual of the markers belonging to cluster no. lcluster. In the plot, the grey value visualises the marker percentage.

Value

list with components

locfreq
vector of individual marker percentages.

locfreqmin
vector of minimum individual marker precentages for each cluster in indclust-clustering (the first value refers to the "noise component", if present).

locfreqmax
vector of maximum individual marker precentages for each cluster in indclust-clustering (the first value refers to the "noise component", if present).

locfreqmean
vector of average individual marker precentages for each cluster in indclust-clustering (the first value refers to the "noise component", if present).

Author(s)

Christian Hennig chrish@stats.ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche

See Also

prabclust

Examples

  options(digits=4)
  data(veronica)
  vei <- prabinit(prabmatrix=veronica[1:50,],distance="jaccard")
  ppv <- prabclust(vei)
  veloci <- prabinit(prabmatrix=veronica[1:50,],rows.are.species=FALSE)
  velociclust <- prabclust(veloci,nnk=0)
  lociplots(ppv,velociclust$clustering,veloci,lcluster=3)

[Package prabclus version 2.2-6 Index]