nb {prabclus}	R Documentation

Neighborhood list for Aegean islands

Description

List of neighboring islands for 34 Aegean islands.

Usage

data(nb)

Format

List with 34 components, all being vetors of integers (or numeric(0) in case of no neighbors) indicating the neighboring islands.

Details

Reads from example data file nb.dat.

Source

B. Hausdorf and C. Hennig (2005) The influence of recent geography, palaeography and climate on the composition of the faune of the central Aegean Islands. Biological Journal of the Linnean Society 84, 785-795.

Examples

data(nb)
# nb <- list()
# for (i in 1:34)
#   nb <- c(nb,list(scan(file="(path/)nb.dat",
#                   skip=i-1,nlines=1)))

[Package prabclus version 2.2-6 Index]