nbtest {prabclus}	R Documentation

Test of neighborhood list

Description

Tests a list of neighboring regions for proper format. Neighborhood is tested for being symmetrical. Causes an error if tests fail.

Usage

nbtest(nblist, n.regions=length(nblist))

Arguments

nblist
A list with a component for every region. The components are vectors of integers indicating neighboring regions. A region without neighbors (e.g., an island) should be assigned a vector numeric(0).

n.regions
Number of regions.

Value

invisible{TRUE}.

Author(s)

Christian Hennig chrish@stats.ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche

See Also

prabinit.

Examples

data(nb)
nbtest(nb)
nb[[1]][1] <- 1
try(nbtest(nb))

[Package prabclus version 2.2-6 Index]