siskiyou {prabclus}	R Documentation

Herbs of the Siskiyou Mountains

Description

Distributions of species of herbs in relation to elevation on quartz diorite in the central Siskiyou Mountains. All values are per mille frequencies in transects (The number of 1 m2 quadrats, among 1000 such quadrats, in which a species was observed, based on 1250 1m2 quadrats in the first 5 transects, and 400 1m2 quadrats in 6. transect). Observed presences in the transect, outside the sampling plots, were coded as 0.2. Rows correspond to species, columns correspond to regions.

Usage

data(siskiyou)

Format

Three objects are generated:

siskiyou

numeric matrix giving the 144*6 abundance values.

siskiyou.nb

neighborhood list for the 6 regions.

siskiyou.groups

integer vector of length 144, giving group memberships for the 144 species.

Details

Reads from example data files LeiMik1.dat, LeiMik1NB.dat, LeiMik1G.dat.

Source

Whittaker, R. H. 1960. Vegetation of the Siskiyou Mountains, Oregon and California. Ecol. Monogr. 30: 279-338 (table 14).

Examples

data(siskiyou)

[Package prabclus version 2.2-6 Index]