tetragonula {prabclus}	R Documentation

Microsatellite genetic data of Tetragonula bees

Description

Genetic data for 236 Tetragonula (Apidae) bees from Australia and Southeast Asia, see Franck et al. (2004). The data give pairs of alleles (codominant markers) for 13 microsatellite loci.

Usage

data(tetragonula)

Format

Two objects are generated:

tetragonula

A data frame with 236 observations and 13 string variables. Strings consist of six digits each. The format is derived from the data format used by the software GENEPOP (Rousset 2010). Alleles have a three digit code, so a value of "258260" on variable V10 means that on locus 10 the two alleles have codes 258 and 260. "000" refers to missing values.

tetragonula.coord

a 236*2 matrix. Coordinates of locations of individuals in decimal format, i.e. the first number is latitude (negative values are South), with minutes and seconds converted to fractions. The second number is longitude (negative values are West).

Details

Reads from example data file Heterotrigona_indoFO.dat.

Source

Franck, P., E. Cameron, G. Good, J.-Y. Rasplus, and B. P. Oldroyd (2004) Nest architecture and genetic differentiation in a species complex of Australian stingless bees. Mol. Ecol. 13, 2317-2331.

Rousset, F. (2010) Genepop 4.0 for Windows and Linux. http://kimura.univ-montp2.fr/~rousset/Genepop.pdf

Examples

data(tetragonula)

[Package prabclus version 2.2-6 Index]