unbuild.charmatrix {prabclus}	R Documentation

Internal: create allele list out of character matrix

Description

Creates a list of lists, such as required by alleledist, from the charmatrix component of an alleleobject.

Usage

   unbuild.charmatrix(charmatrix,n.individuals,n.variables)

Arguments

charmatrix
matrix of characters in which there are two rows for every individual corresponding to the two alleles in every locus (column). Entries are allele codes but missing values are coded as NA.

n.individuals
integer. Number of individuals.

n.variables
integer. Number of loci.

Value

A list of lists. In the "outer" list, there are n.variables lists, one for each locus. In the "inner" list, for every individual there is a vector of two codes (typically characters, see alleleinit) for the two alleles in that locus.

Author(s)

Christian Hennig chrish@stats.ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche

See Also

alleleinit, build.charmatrix

Examples

  data(tetragonula)
  tnb <-
  coord2dist(coordmatrix=tetragonula.coord[1:50,],cut=50,file.format="decimal2",neighbors=TRUE)
  ta <- alleleconvert(strmatrix=tetragonula[1:50,])
  tai <- alleleinit(allelematrix=ta,neighborhood=tnb$nblist,distance="none")
  str(unbuild.charmatrix(tai$charmatrix,50,13)) 

[Package prabclus version 2.2-6 Index]