waterdist {prabclus}	R Documentation

Overwater distances between islands in the Aegean sea

Description

Distance matrix of overwater distances in km between 34 islands in the Aegean sea.

Usage

data(waterdist)

Format

A symmetric 34*34 distance matrix.

Details

Reads from example data file Waterdist.dat, in which there is a 35th column and line with distances to Turkey mainland.

Source

B. Hausdorf and C. Hennig (2005) The influence of recent geography, palaeography and climate on the composition of the faune of the central Aegean Islands. Biological Journal of the Linnean Society 84, 785-795.

Examples

data(waterdist)

[Package prabclus version 2.2-6 Index]